requirements.txt
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
Django==4.1.2
django-cleanup==6.0.0
django-split-settings==1.2.0
djangorestframework==3.14.0
frozenlist==1.3.1
greenlet==1.1.3
idna==3.4
multidict==6.0.2
playwright==1.27.1
pydantic==1.10.2
pyee==8.1.0
pytz==2022.4
sqlparse==0.4.3
typing_extensions==4.4.0
xlrd==1.2.0
yarl==1.8.1
